BatteryOptimization fixes
